Zed2i -> Zed rename, keep Zed2i and zed2i module for backwards compatibility
